+++ js/util.js
@@ -1,17 +1,3 @@
-const TYPES = [
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-]
-
-const RUSSIANTYPES = [
-  'Дворец',
-  'Квартира',
-  'Дом',
-  'Бунгало',
-]
-
 /**
  * Функция, создающая случайное число в указанном интервале с указанной точностью
  *
@@ -19,7 +5,7 @@ const RUSSIANTYPES = [
  * @param {number} max - максимальное значение интервала
  * @param {number} precision - число знаков после запятой
  *
- * @returns {number} случайное число
+ * @returns {number} возвращает случайное число
  */
 const getRandomNumber = function (min = 0, max = 1000, precision = 5) {
   if (min < 0 || max < 0 || precision < 0) {
@@ -41,7 +27,7 @@ const getRandomNumber = function (min = 0, max = 1000, precision = 5) {
  * @param {string} className - класс элемента
  * @param {string} text - текстовое содержимое
  *
- * @returns Html элемент
+ * @returns возвращает созданный Html элемент
  */
 const createElement = (tagName, className, text) => {
   const element = document.createElement(tagName);
@@ -53,31 +39,11 @@ const createElement = (tagName, className, text) => {
 };
 
 /**
- * Функция, возвращающая значение типа жилья на русском языке.
+ * Функция показывает сообщение об ошибке
  *
- * @param {string} currentValue - текущее значение жилья(на английском языке)
- *
- * @returns значение типа жилья на русском языке
+ * * @param {number} alertShowTime  - время показа в милисекундах
  */
-const makeRusType = (currentValue) => {
-  let currentIndex = 0;
-  let russianType = '';
-  TYPES.forEach((value, index) => {
-    if (value === currentValue) {
-      currentIndex = index;
-    }
-  });
-  russianType = RUSSIANTYPES[currentIndex];
-  return russianType;
-};
-
-/**
- * Функция показывающая сообщение об ошибке при загрузке предложений
- *
- * @param {string} message - текстовое содержимое сообщения
- * @param {number} alertShowTime  - время показа в милисекундах
- */
-const showAlert = (message, alertShowTime) => {
+const showAlert = (alertShowTime) => {
   const alertContainer = document.createElement('div');
   alertContainer.style.zIndex = 100;
   alertContainer.style.position = 'absolute';
@@ -89,7 +55,7 @@ const showAlert = (message, alertShowTime) => {
   alertContainer.style.textAlign = 'center';
   alertContainer.style.backgroundColor = 'red';
 
-  alertContainer.textContent = message;
+  alertContainer.textContent = 'Произошла ошибка!';
 
   document.body.append(alertContainer);
 
@@ -98,45 +64,4 @@ const showAlert = (message, alertShowTime) => {
   }, alertShowTime);
 };
 
-/**
- * Функция, показывающая сообщение при успешной отправке формы
- */
-const onSuccessSendFormMessage = () => {
-  const message = document.querySelector('#success').content.querySelector('.success').cloneNode(true);
-  document.addEventListener('keydown',(evt) => {
-    if (evt.keyCode === 27){
-      message.remove();
-    }
-  });
-  document.addEventListener('click',() => {
-    message.remove();
-  });
-  message.style.zIndex = 400;
-  document.querySelector('main').appendChild(message);
-};
-
-/**
- * Функция, показывающая сообщение об ошибке при отправке формы
- */
-const onErrorSendFormMessage = () => {
-  const message = document.querySelector('#error').content.querySelector('.error').cloneNode(true);
-  document.addEventListener('keydown',(evt) => {
-    if (evt.keyCode === 27){
-      message.remove();
-    }
-  });
-  document.addEventListener('click',() => {
-    message.remove();
-  });
-  message.style.zIndex = 400;
-  document.querySelector('main').appendChild(message);
-};
-
-export {
-  getRandomNumber,
-  makeRusType,
-  createElement,
-  showAlert,
-  onSuccessSendFormMessage,
-  onErrorSendFormMessage
-};
+export {getRandomNumber, createElement, showAlert};
